requirements.txt
pya2ldb
PyQt6
